+++ activestorage/lib/active_storage/attached/model.rb
@@ -278,6 +278,8 @@ module ActiveStorage
         end
     end
 
+    attr_writer :attachment_changes # :nodoc:
+
     def attachment_changes # :nodoc:
       @attachment_changes ||= {}
     end
@@ -294,6 +296,15 @@ module ActiveStorage
 
     def reload(*) # :nodoc:
       super.tap { @attachment_changes = nil }
+    end
+
+    def becomes(klass) # :nodoc:
+      super.tap do |became|
+        attachment_changes = @attachment_changes&.each_value do |change|
+          change.record = became
+        end
+        became.attachment_changes = attachment_changes
+      end
     end
   end
 end