+++ activestorage/test/models/attached/one_test.rb
@@ -954,4 +954,10 @@ class ActiveStorage::OneAttachedTest < ActiveSupport::TestCase
       end
     end
   end
+
+  test "preserves attachment changes when using STI" do
+    user = User.new(name: "John", avatar: create_blob(filename: "funky.jpg"))
+    special_user = user.becomes(SpecialUser)
+    assert_predicate special_user.avatar, :attached?
+  end
 end